restructure inference strategies and shared data structures
